feat(web): marketing SEO shell, trust pages, and honest scan framing
- Split home into server page (metadata, JSON-LD) + client shell; share
  copy via marketing-content.ts; Lucide icons replace emoji on features
- Add sitemap/robots, root metadataBase, canonical home OG/twitter fields
- Shared MarketingSiteChrome for docs, trust, and security pages
- Public scan metadata titles/descriptions and OG absolute URLs; CTA copy
  and sample index labels aligned with non-WCAG-claim stance
