+++ apps/web/src/app/docs/api/page.tsx
@@ -1,24 +1,33 @@
 import Link from "next/link";
 import type { Metadata } from "next";
+import { MarketingSiteChrome } from "@/components/marketing/marketing-site-chrome";
+import { pageTitle, PRODUCT_DISPLAY_NAME } from "@/lib/product-brand";
+import { getAppBaseUrl } from "@/lib/site-url";
 
 export const metadata: Metadata = {
-  title: "API & integrations",
-  description:
-    "How to integrate FlexibleAccessible today: MCP server, API keys, and webhooks. A standalone public REST API reference site is not shipped in this deployment.",
+  title: pageTitle("API & integrations"),
+  description: `How to integrate ${PRODUCT_DISPLAY_NAME} today: MCP server, org-scoped API keys, and webhooks. This deployment does not ship a separate public OpenAPI browser.`,
+  alternates: { canonical: "/docs/api" },
+  openGraph: {
+    title: pageTitle("API & integrations"),
+    description: `Integration paths for ${PRODUCT_DISPLAY_NAME}: MCP, API keys, and webhooks.`,
+    url: `${getAppBaseUrl()}/docs/api`,
+    type: "website",
+  },
 };
 
 export default function DocsApiPage() {
   return (
-    <div className="min-h-screen bg-slate-50">
-      <div className="mx-auto max-w-2xl px-6 py-16">
-        <p className="text-sm font-medium text-brand-600">Documentation</p>
+    <MarketingSiteChrome>
+      <div className="mx-auto max-w-2xl px-6 py-section-md">
+        <p className="text-sm font-medium text-brand-700">Documentation</p>
         <h1 className="mt-2 text-3xl font-bold text-slate-900">
           API and integrations
         </h1>
         <p className="mt-4 text-slate-600">
           This product ships integration through authenticated workspace features,
-          not a separate browsable public API reference site. Below is what the
-          repository actually supports today.
+          not a separate browsable public API reference site. Below is what this
+          repository supports today.
         </p>
 
         <ul className="mt-8 space-y-6 text-slate-700">
@@ -26,14 +35,14 @@ export default function DocsApiPage() {
             <h2 className="font-semibold text-slate-900">MCP server</h2>
             <p className="mt-1 text-sm">
               Run{" "}
-              <code className="rounded bg-slate-200 px-1.5 py-0.5 text-xs">
+              <code className="rounded bg-slate-100 px-1.5 py-0.5 text-xs text-slate-800">
                 npx @aros/mcp-server
               </code>{" "}
               for tool-based access. See the package README on npm for options.
             </p>
             <a
               href="https://www.npmjs.com/package/@aros/mcp-server"
-              className="mt-2 inline-block text-sm font-medium text-brand-600 hover:underline"
+              className="mt-2 inline-block text-sm font-medium text-brand-700 hover:underline"
               target="_blank"
               rel="noopener noreferrer"
             >
@@ -41,20 +50,22 @@ export default function DocsApiPage() {
             </a>
           </li>
           <li>
-            <h2 className="font-semibold text-slate-900">Organization API keys</h2>
+            <h2 className="font-semibold text-slate-900">
+              Organization API keys
+            </h2>
             <p className="mt-1 text-sm">
               After you sign in, create and rotate keys under Settings → API keys.
               Usage is org-scoped and subject to plan limits enforced on the server.
             </p>
             <Link
               href="/signup"
-              className="mt-2 inline-block text-sm font-medium text-brand-600 hover:underline"
+              className="mt-2 inline-block text-sm font-medium text-brand-700 hover:underline"
             >
               Create a workspace
             </Link>
           </li>
           <li>
-            <h2 className="font-semibold text-slate-900">Webhooks & CI</h2>
+            <h2 className="font-semibold text-slate-900">Webhooks &amp; CI</h2>
             <p className="mt-1 text-sm">
               Deploy hooks and GitHub Actions integrations are configured in the
               dashboard per site or repository; see in-app settings for the exact
@@ -68,14 +79,7 @@ export default function DocsApiPage() {
           machine-readable contracts, use the MCP package source or ask your
           operator for an export from this deployment.
         </p>
-
-        <Link
-          href="/"
-          className="mt-8 inline-block text-sm font-medium text-slate-600 hover:text-slate-900"
-        >
-          ← Back to home
-        </Link>
       </div>
-    </div>
+    </MarketingSiteChrome>
   );
 }
