+++ apps/web/src/app/robots.ts
@@ -1,0 +1,16 @@
+import type { MetadataRoute } from "next";
+import { getAppBaseUrl } from "@/lib/site-url";
+
+export default function robots(): MetadataRoute.Robots {
+  const base = getAppBaseUrl();
+  const { host } = new URL(base);
+
+  return {
+    rules: {
+      userAgent: "*",
+      disallow: ["/dashboard/", "/settings/"],
+    },
+    sitemap: `${base}/sitemap.xml`,
+    host,
+  };
+}
