+++ apps/web/src/styles/globals.css
@@ -9,6 +9,7 @@
     --color-muted: 100 116 139;
     --color-border: 226 232 240;
     --color-card: 255 255 255;
+    --color-surface-elevated: 255 255 255;
     --color-accent: 13 148 136;
     --color-canvas: 248 250 252;
   }
@@ -64,7 +65,7 @@
   }
 
   .card {
-    @apply rounded-xl border border-slate-200 bg-white p-6 shadow-sm;
+    @apply rounded-xl border border-[rgb(var(--color-border))] bg-[rgb(var(--color-card))] p-6 shadow-sm;
   }
 
   .badge {
